feat: comprehensive topology pipeline overhaul with runtime mode switching
BREAKING CHANGES:
- AIInferenceEngine now requires async initialization and uses real pipeline backend
- Pipeline architecture refactored to support dynamic topology mode switching
- TCSAnalyzer now uses internal state management instead of direct PadGhostState access

CORE PIPELINE IMPROVEMENTS (niodoo_real_integrated/src/pipeline.rs):
- Added TopologyMode enum support with Hybrid and Baseline modes
- Implemented set_topology_mode() for runtime topology switching
- Added initialise_with_mode() and initialise_with_topology() factory methods
- TCS analyzer initialization now conditional on topology mode selection
- Enhanced logging: differentiate between hybrid mode active vs analytic fallback
- Added topology_mode field to PipelineCycle for state tracking
- Improved error handling: graceful fallback when TCS analyzer fails
- Mock mode detection: skip Qdrant collection checks when in mock mode
- Removed unnecessary RwLock import, streamlined Arc usage

AI INFERENCE WRAPPER REWRITE (src/ai_inference.rs):
- Complete redesign: now wraps real niodoo_real_integrated::Pipeline
- Added async initialization methods: new_default(), with_args()
- Topology mode management: getter/setter with pipeline synchronization
- New AIInferenceResult struct with comprehensive metrics:
  * prompt, response, baseline_response for comparison
  * confidence, entropy, rouge quality scores
  * latency_ms for performance tracking
  * topology_mode for state tracking
  * topology_signature with full topological data
  * timings snapshot with per-stage breakdown
- Added StageTimingSnapshot extraction from PipelineCycle
- Integrated with real pad_to_emotional_analysis() conversion
- Proper async/await throughout with Mutex-based concurrency
- Error handling: anyhow::Result propagated correctly

TCS ANALYSIS ENHANCEMENTS (niodoo_real_integrated/src/tcs_analysis.rs):
- Introduced TCSState struct for internal state management
- Added TCSHandle type alias for Arc<Mutex<TCSState>>
- pad_to_points() now accepts TCSState instead of PadGhostState
- pad_to_knot_diagram() refactored for cleaner state handling
- State initialization: converts PadGhostState to TCSState with proper type casting
- Improved thread safety: all state access through Mutex guards
- Added unit tests for TCS delta computation
- Better separation of concerns: analyzer manages its own state representation

COMPASS & EMBEDDING IMPROVEMENTS:
- Compass: Updated error handling and dependency management
- Embedding: Enhanced vector processing with better error propagation

FEDERATED & LEARNING UPDATES:
- Federated: Streamlined network topology for distributed operations
- Learning: Reduced complexity, removed 62 lines of redundant code
- Better separation between learning logic and topology computation

TOKEN MANAGER & MAIN:
- Token manager: Refactored for cleaner state transitions (29 lines simplified)
- Main: Reduced complexity with better separation of concerns (96 lines removed)
- Removed unnecessary initialization steps

TESTING INFRASTRUCTURE:
- Integration tests: Added 16 new assertions for topology verification
- Better test isolation and clearer failure messages
- Performance benchmark script updated with new timing modes

PERFORMANCE & COGNITIVE IMPROVEMENTS:
- AI inference: Complete rewrite eliminates mock data, uses real topology
- Consciousness: Streamlined to 60 fewer lines with better algorithms
- Emotional coder: Added 39 lines of real emotional state processing
- Empathy: Enhanced emotional state tracking (13 lines added)
- Multi-layer query: Improved memory layer transitions (14 lines updated)
- Token promotion simulation: Better integration with topology pipeline

BUILD & DEPENDENCIES:
- Added qdrant_client::prelude import in pipeline
- Updated build.rs with proto configuration
- Cargo.toml: Streamlined dependency management

METRICS & OBSERVABILITY:
- All topology metrics now properly recorded
- Stage timings tracked across entire pipeline
- Better error logging with context preservation
- Topology signatures exposed in AI inference results

PERFORMANCE BENCHMARKS:
- Updated run_performance_benchmarks.sh with new timing infrastructure
- Support for multiple topology modes in benchmark runs
- Better comparison between baseline and hybrid modes

This commit represents a major architectural improvement, moving from mock/stub
implementations to production-ready topology-aware AI inference with full support
for runtime mode switching and comprehensive metrics collection.

+++ tcs-pipeline/src/lib.rs
@@ -178,8 +178,8 @@ impl TCSOrchestrator {
             }
             if feature.persistence() > self.config.persistence_event_threshold {
                 let pf = PersistentFeature {
-                    birth: feature.birth,
-                    death: feature.death,
+                    birth: feature.birth as f64,
+                    death: feature.death as f64,
                     dimension: feature.dimension,
                 };
                 self.state.register_feature(pf.clone());
